Move upgrades to their own folder
